Added Activity enum and fixed sale button error

+++ InventoryManagementApp/User Controls/Orders_UC.cs
@@ -20,6 +20,7 @@ public partial class Orders_UC : UserControl
         toDate.Enabled = false;
         fromDate.Enabled = false;
         dateCheck.Enabled = false;
+        ordersData.RowPrePaint += ordersData_RowPrePaint;
         ordersData.DataSource = _db.Sales.Select(o => new
         {
             კოდი = o.Id,
@@ -30,8 +31,35 @@ public partial class Orders_UC : UserControl
             ლოკაცია = o.Location,
             გადახდის_ტიპი = o.PaymentMethod,
             შეკვეთის_ადგილი = o.PaymentArea,
+            მდგომარეობა = o.Activity,
             თარიღი = o.Date
         }).ToList();
+
+
+    }
+
+    private void ordersData_RowPrePaint(object sender, DataGridViewRowPrePaintEventArgs e)
+    {
+        DataGridViewRow row = ordersData.Rows[e.RowIndex];
+
+        if (row.Cells["მდგომარეობა"] is DataGridViewTextBoxCell cell)
+        {
+            string state = cell.Value.ToString();
+
+            if (state == "მზადდება")
+            {
+                cell.Style.ForeColor = Color.Red;
+            }
+            else if (state == "გზაშია")
+            {
+                cell.Style.ForeColor = Color.Orange;
+            }
+            else if (state == "მიწოდებულია")
+            {
+                cell.Style.ForeColor = Color.Green;
+            }
+        }
+
     }
 
     private void edit_Btn_Click(object sender, EventArgs e)
@@ -46,46 +74,50 @@ public partial class Orders_UC : UserControl
 
     private void delete_Btn_Click(object sender, EventArgs e)
     {
-        var row = ordersData.CurrentRow;
-        var orderIndex = ordersData.CurrentRow.Cells["კოდი"].Value.ToString();
-        var orderId = orderIndex;
-        var order = _db.Sales.FirstOrDefault(p => p.Id.ToString() == orderId);
-
-        order.DeleteOrder();
-        order.DeleteTime = DateTime.Now;
-
-        var response = _db.SaveChanges();
-
-        if(response > 0)
+        if (MessageBox.Show("დარწმუნებული ხართ, რომ გსურთ წაშლა?", "ფრთხილად წაშლამდე!"
+                           , MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
         {
-            MessageBox.Show("შეკვეთა წარმატებით წაიშალა!",
-                              "შეკვეთის წაშლა",
-                              MessageBoxButtons.OK,
-                              MessageBoxIcon.Information);
+            var row = ordersData.CurrentRow;
+            var orderIndex = ordersData.CurrentRow.Cells["კოდი"].Value.ToString();
+            var orderId = orderIndex;
+            var order = _db.Sales.FirstOrDefault(p => p.Id.ToString() == orderId);
 
-            ordersData.DataSource = _db.Sales.Select(o => new
+            order.DeleteOrder();
+            order.DeleteTime = DateTime.Now;
+
+            var response = _db.SaveChanges();
+
+            if (response > 0)
             {
-                კოდი = o.Id,
-                პროდუქტი = o.Product.Name,
-                კატეგორია = o.Product.Category.Name,
-                ბრენდი = o.Product.Brand.Name,
-                რაოდენობა = o.Amount,
-                ლოკაცია = o.Location,
-                გადახდის_ტიპი = o.PaymentMethod,
-                შეკვეთის_ადგილი = o.PaymentArea,
-                თარიღი = o.Date
-            }).ToList();
+                MessageBox.Show("შეკვეთა წარმატებით წაიშალა!",
+                                  "შეკვეთის წაშლა",
+                                  MessageBoxButtons.OK,
+                                  MessageBoxIcon.Information);
+
+                ordersData.DataSource = _db.Sales.Select(o => new
+                {
+                    კოდი = o.Id,
+                    პროდუქტი = o.Product.Name,
+                    კატეგორია = o.Product.Category.Name,
+                    ბრენდი = o.Product.Brand.Name,
+                    რაოდენობა = o.Amount,
+                    ლოკაცია = o.Location,
+                    გადახდის_ტიპი = o.PaymentMethod,
+                    შეკვეთის_ადგილი = o.PaymentArea,
+                    მდგომარეობა = o.Activity,
+                    თარიღი = o.Date
+                }).ToList();
+
+            }
+            else
+            {
+                MessageBox.Show("შეკვეთის წაშლა ვერ მოხერხდა, თავიდან სცადეთ!",
+                                   "შეცდომა წაშლისას",
+                                   MessageBoxButtons.OK,
+                                   MessageBoxIcon.Warning);
+            }
 
         }
-        else
-        {
-            MessageBox.Show("შეკვეთის წაშლა ვერ მოხერხდა, თავიდან სცადეთ!",
-                               "შეცდომა წაშლისას",
-                               MessageBoxButtons.OK,
-                               MessageBoxIcon.Warning);
-        }
-
-
     }
 
     private void searchButton_Click(object sender, EventArgs e)
@@ -105,6 +137,7 @@ public partial class Orders_UC : UserControl
                 ლოკაცია = o.Location,
                 გადახდის_ტიპი = o.PaymentMethod,
                 შეკვეთის_ადგილი = o.PaymentArea,
+                მდგომარეობა = o.Activity,
                 თარიღი = o.Date
             }).ToList();
             ordersData.DataSource = onlySale;
@@ -122,6 +155,7 @@ public partial class Orders_UC : UserControl
                 ლოკაცია = o.Location,
                 გადახდის_ტიპი = o.PaymentMethod,
                 შეკვეთის_ადგილი = o.PaymentArea,
+                მდგომარეობა = o.Activity,
                 თარიღი = o.Date
             }).ToList();
             ordersData.DataSource = sales;
@@ -161,6 +195,7 @@ public partial class Orders_UC : UserControl
                 ლოკაცია = o.Location,
                 გადახდის_ტიპი = o.PaymentMethod,
                 შეკვეთის_ადგილი = o.PaymentArea,
+                მდგომარეობა = o.Activity,
                 თარიღი = o.Date
             }).ToList();
         }
@@ -183,6 +218,7 @@ public partial class Orders_UC : UserControl
                                       ლოკაცია = s.Location,
                                       გადახდის_ტიპი = s.PaymentMethod,
                                       შეკვეთის_ადგილი = s.PaymentArea,
+                                      მდგომარეობა = s.Activity,
                                       თარიღი = s.Date
                                   }).ToList();
 
@@ -246,8 +282,8 @@ public partial class Orders_UC : UserControl
     {
         var from_date = fromDate.Value.Date;
         var to_date = toDate.Value.Date.AddDays(1);
-        if (dateCheck.Checked) 
-        {          
+        if (dateCheck.Checked)
+        {
             var onlyDate = _db.Sales.Where(d => d.Date == from_date).Select(d => d.Product);
             var onlySale = _db.Sales.Where(o => onlyDate.Contains(o.Product) && o.Date == from_date).Select(o => new
             {
@@ -259,6 +295,7 @@ public partial class Orders_UC : UserControl
                 ლოკაცია = o.Location,
                 გადახდის_ტიპი = o.PaymentMethod,
                 შეკვეთის_ადგილი = o.PaymentArea,
+                მდგომარეობა = o.Activity,
                 თარიღი = o.Date
             }).ToList();
             ordersData.DataSource = onlySale;
@@ -276,12 +313,13 @@ public partial class Orders_UC : UserControl
                 ლოკაცია = o.Location,
                 გადახდის_ტიპი = o.PaymentMethod,
                 შეკვეთის_ადგილი = o.PaymentArea,
+                მდგომარეობა = o.Activity,
                 თარიღი = o.Date
             }).ToList();
             ordersData.DataSource = sales;
         }
-        
-        
+
+
     }
 
     private void toDate_ValueChanged(object sender, EventArgs e)
@@ -299,8 +337,10 @@ public partial class Orders_UC : UserControl
             ლოკაცია = o.Location,
             გადახდის_ტიპი = o.PaymentMethod,
             შეკვეთის_ადგილი = o.PaymentArea,
+            მდგომარეობა = o.Activity,
             თარიღი = o.Date
         }).ToList();
         ordersData.DataSource = sales;
     }
+
 }
